+++ lib/app/server/middleware/typegraphql-interfaces.d.ts
@@ -1,0 +1,3 @@
+import { MiddlewareFn, MiddlewareInterface } from 'type-graphql';
+export declare type MiddlewareClass<TContext = {}> = new (...args: any[]) => MiddlewareInterface<TContext>;
+export declare type Middleware<TContext = {}> = MiddlewareFn<TContext> | MiddlewareClass<TContext>;