+++ lib/app/server/plugins/good-rfi.d.ts
@@ -1,0 +1,7 @@
+import * as Hapi from '@hapi/hapi';
+export declare const goodRfi: {
+    name: string;
+    register(server: Hapi.Server, server_config: {
+        logging: object;
+    }): Promise<void>;
+};