+++ lib/app/server/plugins/index.d.ts
@@ -1,0 +1,2 @@
+export * from './good-rfi';
+export * from './apollo-error.plugin';