+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+  * Centralized `initServer` logic
+### Removed
+  * `iterable-model` is no longer a peer
+### Changed
+  * Extract convict config schemas to separate package, `@rumbleship/config`
+### Fixed
+### Deprecated
+### Security
+
+
 ## [9.1.0] -- 2020-06-26
 
 ### Added